Migrate csetting
